psdev-1120: Fix pylint in infini_export
Use a var, pool_size, in test_create_resources playbook.  Set in
ibox_vars files.  This needs to be a var since it needs to be at least 1TB in a
real IBOX, but has to be very small, 10GB, in a virtual IBOX.

DCO 1.1 Signed-off-by: David Ohlemacher <dohlemacher@infinidat.com>